Use integer types for offsets, ids, sizes and versions
File offsets, node ids, sizes and versions traveled as double
throughout the B+ tree, R-tree and text log, inviting float-equality
and (size_t)offset-truncation bugs. They are now uint64_t (offsets,
ids, versions) and int64_t (sizes, counts) internally; double remains
only for keys, coordinates and scores.

Conversions are confined to two seams: the wire, where POINTER
payloads are already uint64 LE and ids/sizes/versions decode through a
new read_u64 that rejects fractional or negative values from hostile
files, and the *_wasm.c glue, which converts to the doubles JS expects
(JS-visible signatures unchanged). The rtree compaction ptrmap keys on
native uint64_t offsets instead of double bit patterns.

On-wire bytes are unchanged: all JS<->WASM interop tests pass and
performance is flat within noise.

+++ include/bplustree.h
@@ -47,10 +47,11 @@ bpt *bpt_open(const bj_io *io);
 /* Free a tree and all its buffers (does not touch the file). Safe on NULL. */
 void bpt_free(bpt *t);
 
-/* Accessors (mirror the JS metadata fields). */
-double         bpt_size(const bpt *t);
-double         bpt_root(const bpt *t);
-double         bpt_next_id(const bpt *t);
+/* Accessors (mirror the JS metadata fields; the WASM glue converts to the
+ * doubles JS expects). */
+int64_t        bpt_size(const bpt *t);
+uint64_t       bpt_root(const bpt *t);
+uint64_t       bpt_next_id(const bpt *t);
 int            bpt_order(const bpt *t);
 
 /* Insert/update. `val`/`val_len` is one pre-encoded binjson value (opaque). */
